# OpenAPI-Specification.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Invite API",
    "version": "v1"
  },
  "paths": {
    "/invite/api/redeemed-referrals": {
      "get": {
        "tags": [
          "RedeemedReferrals"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReferralRedemption"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReferralRedemption"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReferralRedemption"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/invite/api/redemptions/{id}": {
      "get": {
        "tags": [
          "Redemptions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RedemptionRequest"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RedemptionRequest"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RedemptionRequest"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/invite/api/redemptions": {
      "post": {
        "tags": [
          "Redemptions"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RedemptionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RedemptionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RedemptionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/invite/api/referral-link": {
      "get": {
        "tags": [
          "ReferralLinks"
        ],
        "parameters": [
          {
            "name": "source",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ReferralSourceEnum"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReferralLink"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReferralLink"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReferralLink"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/invite/api/referrals": {
      "get": {
        "tags": [
          "Referrals"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReferralResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReferralResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReferralResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Referrals"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReferralRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReferralRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ReferralRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/invite/api/referrals/{id}": {
      "get": {
        "tags": [
          "Referrals"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReferralResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReferralResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReferralResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "RedemptionRequest": {
        "required": [
          "referralCode",
          "referralSource"
        ],
        "type": "object",
        "properties": {
          "referralCode": {
            "maxLength": 6,
            "minLength": 6,
            "pattern": "^[a-zA-Z0-9]+$",
            "type": "string"
          },
          "referralSource": {
            "$ref": "#/components/schemas/ReferralSourceEnum"
          }
        },
        "additionalProperties": false
      },
      "ReferralLink": {
        "required": [
          "referralLink"
        ],
        "type": "object",
        "properties": {
          "referralLink": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "ReferralRedemption": {
        "required": [
          "firstName",
          "lastInitial",
          "redeemedAt",
          "referralCode",
          "referralStatus"
        ],
        "type": "object",
        "properties": {
          "firstName": {
            "minLength": 1,
            "type": "string"
          },
          "lastInitial": {
            "minLength": 1,
            "type": "string"
          },
          "referralCode": {
            "maxLength": 6,
            "minLength": 6,
            "pattern": "^[a-zA-Z0-9]+$",
            "type": "string"
          },
          "referralStatus": {
            "$ref": "#/components/schemas/ReferralStatusEnum"
          },
          "redeemedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "ReferralRequest": {
        "required": [
          "referralCode",
          "referralSource"
        ],
        "type": "object",
        "properties": {
          "referralCode": {
            "maxLength": 6,
            "minLength": 6,
            "pattern": "^[a-zA-Z0-9]+$",
            "type": "string"
          },
          "referralSource": {
            "$ref": "#/components/schemas/ReferralSourceEnum"
          }
        },
        "additionalProperties": false
      },
      "ReferralResponse": {
        "required": [
          "createdAt",
          "referralCode",
          "referralId",
          "referralSource"
        ],
        "type": "object",
        "properties": {
          "referralId": {
            "type": "string",
            "format": "uuid"
          },
          "referralCode": {
            "maxLength": 6,
            "minLength": 6,
            "pattern": "^[a-zA-Z0-9]+$",
            "type": "string"
          },
          "referralSource": {
            "$ref": "#/components/schemas/ReferralSourceEnum"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "ReferralSourceEnum": {
        "enum": [
          "Android",
          "iOS",
          "Chrome",
          "Edge",
          "Firefox"
        ],
        "type": "string"
      },
      "ReferralStatusEnum": {
        "enum": [
          "Complete"
        ],
        "type": "string"
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme.",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}